requirements.txt
backcall==0.2.0
base58==2.1.1
blinker==1.4
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
charset-normalizer==2.0.10
dashboard==0.0.6
databases==0.5.3
debugpy==1.5.1
decorator==5.1.0
dj-database-url==0.5.0
Django
django-allauth
django-appconf
django-autoslug==1.9.8
django-bootstrap3
django-crispy-forms==1.13.0
django-dash==0.6.1
django-friendship==1.9.4
django-heroku
django-messages
django-nine
django-notification==1.3.3
django-registration
django-smart-selects==1.5.9
django-tinymce==3.4.0
django-user-accounts==3.2.0
djangorestframework==3.13.1
google-auth==2.6.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.2
grpcio==1.44.0
gunicorn==20.1.0
h11==0.12.0
h5py==3.6.0
humanize==4.0.0
hyperlink==21.0.0
Markdown==3.3.6
matplotlib
MarkupSafe==2.0.1
mysql==0.0.3
mysqlclient==2.1.0
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
notification==0.2.1
numpy==1.21.5
oauthlib==3.2.0
opt-einsum==3.3.0
orm==0.2.0.dev1
pandas==1.3.5
PyMySQL
pyOpenSSL==22.0.0
pyparsing==3.0.6
pyrsistent==0.18.1
PyQt5
python-dateutil==2.8.2
python-multipart==0.0.5
python3-openid==3.2.0
requests==2.27.1
requests-oauthlib==1.3.1
virtualenv==20.13.3
scikit-learn==1.0.2
